// Engenharia_Software/Frontend/src/app/interfaces/IMedico.ts
export interface IMedico{
    Medico_ID: string;
    Nome: string;
    CPF: string;
    CRM: string;
    Cidade: string;
    Estado: string;
    Logradouro: string;
    Logradouro_Numero: string;
    CEP: string;
    Telefone: string;
    Celular: string;
    Email: string;
    Descricao: string;
    Sexo: string;
    Data_Nascimento: string;
    Senha: string;
  }