// Engenharia_Software/Frontend/src/app/interfaces/IPaciente.ts
export interface IPaciente{
    Paciente_ID: string;
    Nome: string;
    Data_Nascimento: string;
    CPF: string;
    Cidade: string;
    Estado: string;
    Logradouro: string;
    Logradouro_Numero: number;
    CEP: string;
    Telefone: string;
    Celular: string;
    Email: string;
    Senha: string;
  }